% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_features.R
\name{calculate_hydrological_variables}
\alias{calculate_hydrological_variables}
\title{Calcul des variables hydrologiques}
\usage{
calculate_hydrological_variables(dem, rivers = NULL, watershed = NULL)
}
\arguments{
\item{dem}{SpatRaster avec altitude et pente (issu de load_dem)}

\item{rivers}{Objet sf des cours d'eau (optionnel)}

\item{watershed}{Objet sf des bassins versants (optionnel)}
}
\value{
Une liste avec le raster hydrologique et les variables par bassin
}
\description{
Calcul des variables hydrologiques
}
\examples{
\dontrun{
hydro <- calculate_hydrological_variables(dem, rivers, watershed)
}
}
