% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environmental_features.R
\name{extract_environmental_features}
\alias{extract_environmental_features}
\title{Extraction des variables environnementales aux stations}
\usage{
extract_environmental_features(water_quality, landuse, hydro, rainfall = NULL)
}
\arguments{
\item{water_quality}{Objet sf des stations qualité eau}

\item{landuse}{SpatRaster d'occupation du sol}

\item{hydro}{Liste issue de calculate_hydrological_variables}

\item{rainfall}{SpatRaster de pluie (optionnel)}
}
\value{
Un dataframe avec toutes les variables par station
}
\description{
Extraction des variables environnementales aux stations
}
\examples{
\dontrun{
features <- extract_environmental_features(stations, landuse, hydro)
}
}
