% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_report}
\alias{generate_report}
\title{Génération d'un rapport HTML ou PDF automatique}
\usage{
generate_report(
  eval_result,
  summary_df,
  cartes,
  data_info,
  output_file = "rapport_waterQualSDM.html"
)
}
\arguments{
\item{eval_result}{Liste issue de evaluate_model}

\item{summary_df}{Dataframe issu de generate_recommendations}

\item{cartes}{Liste issue de plot_water_quality_map}

\item{data_info}{Liste avec infos sur les données utilisées}

\item{output_file}{Chemin du rapport (défaut: "rapport_waterQualSDM.html")}
}
\value{
Le chemin du rapport généré
}
\description{
Génération d'un rapport HTML ou PDF automatique
}
\examples{
\dontrun{
generate_report(eval_result, summary_df, cartes, data_info)
}
}
