% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_watershed}
\alias{import_watershed}
\title{Import des bassins versants}
\usage{
import_watershed(file, crs = 4326)
}
\arguments{
\item{file}{Chemin vers le shapefile ou GeoJSON}

\item{crs}{Système de coordonnées cible (défaut: 4326)}
}
\value{
Un objet sf des bassins versants
}
\description{
Import des bassins versants
}
\examples{
\dontrun{
ws <- import_watershed("bassins.shp")
}
}
