% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{load_dem}
\alias{load_dem}
\title{Import du Modèle Numérique de Terrain}
\usage{
load_dem(file, watershed = NULL, target_res = NULL)
}
\arguments{
\item{file}{Chemin vers le fichier raster (GeoTIFF)}

\item{watershed}{Objet sf pour découper le raster (optionnel)}

\item{target_res}{Résolution cible en mètres pour harmonisation (défaut: NULL)}
}
\value{
Un objet SpatRaster avec altitude et pente
}
\description{
Import du Modèle Numérique de Terrain
}
\examples{
\dontrun{
dem <- load_dem("mnt.tif", target_res = 100)
}
}
