% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_soc_distribution}
\alias{plot_soc_distribution}
\title{Distribution du SOC par classe d'occupation du sol}
\usage{
plot_soc_distribution(water_quality, soc_col = "soc", landuse_col = "landuse")
}
\arguments{
\item{water_quality}{Dataframe des stations de mesure}

\item{soc_col}{Nom de la colonne SOC (défaut: "soc")}

\item{landuse_col}{Nom de la colonne occupation du sol (défaut: "landuse")}
}
\value{
Un objet ggplot2
}
\description{
Distribution du SOC par classe d'occupation du sol
}
\examples{
\dontrun{
data(water_quality)
plot_soc_distribution(water_quality)
}
}
