% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_water_quality_map}
\alias{plot_water_quality_map}
\title{Cartographie de la qualité de l'eau}
\usage{
plot_water_quality_map(
  nitrates_raster,
  vuln_raster,
  watershed,
  output_file = NULL
)
}
\arguments{
\item{nitrates_raster}{SpatRaster des nitrates prédits}

\item{vuln_raster}{SpatRaster de vulnérabilité}

\item{watershed}{Objet sf des bassins versants}

\item{output_file}{Chemin pour exporter la carte PNG ou PDF (optionnel)}
}
\value{
Une liste de 3 cartes ggplot2
}
\description{
Cartographie de la qualité de l'eau
}
\examples{
\dontrun{
cartes <- plot_water_quality_map(nitrates_map, vuln, watershed)
}
}
