% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rivers_sf}
\alias{rivers_sf}
\title{Cours d'eau simulés}
\format{
Un objet sf avec 2 lignes et 3 variables :
\describe{
  \item{river_id}{Identifiant du cours d'eau}
  \item{river_name}{Nom du cours d'eau}
  \item{geometry}{Géométrie de la ligne}
}
}
\source{
Données simulées
}
\usage{
rivers_sf
}
\description{
Un objet sf contenant 2 cours d'eau fictifs.
}
\keyword{datasets}
