% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{summarize_watersheds}
\alias{summarize_watersheds}
\title{Résumé statistique par bassin versant}
\usage{
summarize_watersheds(watershed, nitrates_raster, vuln_raster, landuse)
}
\arguments{
\item{watershed}{Objet sf des bassins versants}

\item{nitrates_raster}{SpatRaster de la variable qualité prédite (SOC ou nitrates)}

\item{vuln_raster}{SpatRaster de vulnérabilité}

\item{landuse}{SpatRaster d'occupation du sol}
}
\value{
Un dataframe avec les statistiques par bassin
}
\description{
Résumé statistique par bassin versant
}
\examples{
\dontrun{
summary <- summarize_watersheds(watershed, soc_map, vuln)
}
}
