% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_sf}
\alias{watershed_sf}
\title{Régions administratives du Maroc}
\format{
Un objet sf avec 16 lignes et 2 variables :
\describe{
  \item{NAME_1}{Nom de la région}
  \item{geometry}{Géométrie du polygone}
}
}
\source{
GADM 4.1 (https://gadm.org)
}
\usage{
watershed_sf
}
\description{
Un objet sf contenant les régions administratives du Maroc
et le Sahara Occidental (source: GADM 4.1).
}
\keyword{datasets}
