% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{calculate_vulnerability_index}
\alias{calculate_vulnerability_index}
\title{Calcul de l'indice de vulnérabilité}
\usage{
calculate_vulnerability_index(
  nitrates_raster,
  landuse,
  hydro,
  output_file = NULL
)
}
\arguments{
\item{nitrates_raster}{SpatRaster des nitrates prédits}

\item{landuse}{SpatRaster d'occupation du sol}

\item{hydro}{SpatRaster des variables hydrologiques}

\item{output_file}{Chemin pour exporter le GeoTIFF (optionnel)}
}
\value{
Un SpatRaster de l'indice de vulnérabilité (1=faible, 2=moyen, 3=élevé)
}
\description{
Calcul de l'indice de vulnérabilité
}
\examples{
\dontrun{
vuln <- calculate_vulnerability_index(nitrates_map, landuse, hydro)
}
}
