% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Génération de recommandations de gestion}
\usage{
generate_recommendations(summary_df)
}
\arguments{
\item{summary_df}{Dataframe issu de summarize_watersheds}
}
\value{
Un dataframe avec les recommandations par bassin
}
\description{
Génération de recommandations de gestion
}
\examples{
\dontrun{
reco <- generate_recommendations(summary_df)
}
}
