% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_water_quality}
\alias{predict_water_quality}
\title{Prédiction spatiale de la qualité du sol}
\usage{
predict_water_quality(model_result, landuse, hydro, output_file = NULL)
}
\arguments{
\item{model_result}{Liste issue de train_rf_model}

\item{landuse}{SpatRaster d'occupation du sol}

\item{hydro}{SpatRaster des variables environnementales}

\item{output_file}{Chemin pour exporter le GeoTIFF (optionnel)}
}
\value{
Un SpatRaster de la variable qualité prédite (SOC en g/kg)
}
\description{
Prédiction spatiale de la qualité du sol
}
\examples{
\dontrun{
soc_map <- predict_water_quality(model_result, landuse, hydro)
}
}
