% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{train_rf_model}
\alias{train_rf_model}
\title{Entraînement du modèle Random Forest}
\usage{
train_rf_model(data_prep, target = "soc", ntrees = 500, tune = TRUE)
}
\arguments{
\item{data_prep}{Liste issue de preprocess_data}

\item{target}{Nom de la colonne cible (défaut: "soc")}

\item{ntrees}{Nombre d'arbres (défaut: 500)}

\item{tune}{Logical, effectuer un tuning simple (défaut: TRUE)}
}
\value{
Une liste avec le modèle et l'importance des variables
}
\description{
Entraînement du modèle Random Forest
}
\examples{
\dontrun{
model <- train_rf_model(data_prep)
}
}
