% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{watershed_sf}
\alias{watershed_sf}
\title{Bassins versants simulés}
\format{
Un objet sf avec 5 lignes et 3 variables :
\describe{
  \item{basin_id}{Identifiant du bassin}
  \item{basin_name}{Nom du bassin}
  \item{geometry}{Géométrie du polygone}
}
}
\source{
Données simulées
}
\usage{
watershed_sf
}
\description{
Un objet sf contenant 5 bassins versants fictifs.
}
\keyword{datasets}
